add dyndns-like registration of the LAN-side client daemon
Use `python -m flancer.server add-dyndns gw.sf.example.com` on the server
side to generate a code, then type that code into `python -m flancer.client
add-dyndns` on the client side to accept it.

After that, the server will look at the client's connection to deduce its IP
address, and will add an A or AAAA for the configured
hostname (`gw.sf.example.com`) to point at that address. This connection will
be checked every few minutes, so if/when the client's DHCP-provided address
changes, the connection will be re-established and the DNS record will be
updated with the new value.