server.tap: set Tub keepalive/timeouts just like on client
Previously, the client (but not the server) would ping after 60 seconds of
idle time, and disconnects if it doesn't see a response after 5 minutes. If
the client's IP address changed on it (e.g. your DHCP lease from your ISP was
not renewed, and the replacement lease moved you to a different address),
then the client's PING should trigger an ICMP error (severing the TCP
connection), or at least the PING wouldn't make it to the server (since it's
now arriving with the wrong source address), so the server won't respond, and
eventually the client will hit their disconnectTimeout and drop the
connection. Once the connection is dropped, the foolscap Reconnector will
establish a new one (from the new IP address), and the dyndns record should
be updated to the new value.

If the client disappears entirely, the server might wait a very long
time (possibly forever?) for the TCP connection to go away: the default
settings are usually very optimistic. By turning on keepalives and timeouts
on the server too, a lost client should be more visible. Running `lsof` on
the server should show the lack of a client connection.

Also, in the future, we might remove the dyndns record if we ever lose the
client connection.